moved bookingChoice back to pages
