+++ client/src/pages/bookingChoice.jsx
